+++ clients/margin-trading/examples/rest-api/TradeApi/queryPreventedMatches.ts
@@ -1,0 +1,26 @@
+import { MarginTrading, MARGIN_TRADING_REST_API_PROD_URL } from '../../../src';
+
+const configurationRestAPI = {
+    apiKey: process.env.API_KEY ?? '',
+    apiSecret: process.env.API_SECRET ?? '',
+    basePath: process.env.BASE_PATH ?? MARGIN_TRADING_REST_API_PROD_URL,
+};
+const client = new MarginTrading({ configurationRestAPI });
+
+async function queryPreventedMatches() {
+    try {
+        const response = await client.restAPI.queryPreventedMatches({
+            symbol: 'symbol_example',
+        });
+
+        const rateLimits = response.rateLimits!;
+        console.log('queryPreventedMatches() rate limits:', rateLimits);
+
+        const data = await response.data();
+        console.log('queryPreventedMatches() response:', data);
+    } catch (error) {
+        console.error('queryPreventedMatches() error:', error);
+    }
+}
+
+queryPreventedMatches();